More test splitting for travis.
