+++ tests/CMakeLists.txt
@@ -103,6 +103,7 @@ ADD_PIRANHA_TESTCASE(series_04)
 ADD_PIRANHA_TESTCASE(series_05)
 ADD_PIRANHA_TESTCASE(series_06)
 ADD_PIRANHA_TESTCASE(series_07)
+ADD_PIRANHA_TESTCASE(series_08)
 ADD_PIRANHA_TESTCASE(settings)
 ADD_PIRANHA_TESTCASE(small_vector_01)
 ADD_PIRANHA_TESTCASE(small_vector_02)
